+++ docs/.vuepress/config.js
@@ -19,27 +19,39 @@ module.exports = {
         editLinks: true, //  "Edit this page" at the bottom of each page
         lastUpdated: 'Last Updated', //  "Last Updated" at the bottom of each page
         repo: 'nuwave/lighthouse', //  Github repo
-        docsDir: 'docs/', //  Github repo docs folder    
+        docsDir: 'docs/', //  Github repo docs folder
         versions: {
             latest: versioning.versions.latest,
             selected: versioning.versions.latest,
             all: versioning.versions.all
         },
-        search: false,
         nav: [
             {
-                text: `Docs`,
+                text: 'Docs',
                 items: versioning.linksFor('getting-started/installation.md') // TODO create custom component
             },
             {
+                text: 'Tutorial',
+                link: '/tutorial/'
+            },
+            {
                 text: 'Resources',
-                link: '/pages/resources.md'
+                link: '/resources/'
             },
             {
                 text: 'Users',
-                link: '/pages/users.md'
+                link: '/users/'
             }
         ],
         sidebar: versioning.sidebars
-    }
+    },
+    plugins: [
+        ['@vuepress/back-to-top', true],
+        ['@vuepress/medium-zoom', true],
+        ['@vuepress/search', {
+            searchMaxSuggestions: 10,
+            // Only search the latest version, e.g. 4.3, otherwise many duplicates will show up
+            test: `/${versioning.versions.latest.replace('.', '\\.')}/`
+        }]
+    ]
 }