+++ src/Schema/Factories/FieldFactory.php
@@ -23,6 +23,7 @@ use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
 use Nuwave\Lighthouse\Support\Contracts\HasArgumentPath;
 use Nuwave\Lighthouse\Support\Contracts\ProvidesResolver;
 use Nuwave\Lighthouse\Support\Traits\HasResolverArguments;
+use Nuwave\Lighthouse\Execution\Arguments\SpreadMiddleware;
 use Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective;
 use Nuwave\Lighthouse\Support\Contracts\ArgDirectiveForArray;
 use Nuwave\Lighthouse\Support\Contracts\ArgTransformerDirective;
@@ -97,13 +98,6 @@ class FieldFactory
      * @var mixed[]
      */
     protected $handleArgDirectivesSnapshots = [];
-
-    /**
-     * Arg paths to spread out.
-     *
-     * @var array[]
-     */
-    protected $pathsToSpread = [];
 
     /**
      * @param  \Nuwave\Lighthouse\Schema\Factories\DirectiveFactory  $directiveFactory
@@ -194,36 +188,15 @@ class FieldFactory
                 // Now that we are finished with all argument based validation,
                 // we flush the validation error buffer
                 $this->flushValidationErrorBuffer();
-//
-//                // Apply the argument spreadings after we are finished with all
-//                // the other argument handling
-//                foreach ($this->pathsToSpread as $argumentPath) {
-//                    $inputValues = $this->argValue($argumentPath);
-//
-//                    // If no input is given, there is nothing to spread
-//                    if (! $inputValues) {
-//                        continue;
-//                    }
-//
-//                    // We remove the value from where it was defined before
-//                    $this->unsetArgValue($argumentPath);
-//
-//                    // The last part of the path is the name of the input value,
-//                    // the exact thing we want to remove
-//                    array_pop($argumentPath);
-//
-//                    foreach ($inputValues as $key => $value) {
-//                        $this->setArgValue(
-//                            array_merge($argumentPath, [$key]),
-//                            $value
-//                        );
-//                    }
-//                }
 
                 // The final resolver can access the builder through the ResolveInfo
                 $this->resolveInfo->builder = $this->builder;
 
-                return $resolverWithMiddleware($this->root, $this->args, $this->context, $this->resolveInfo);
+                /** @var \Nuwave\Lighthouse\Execution\Arguments\SpreadMiddleware $spreadMiddleware */
+                $spreadMiddleware = app(SpreadMiddleware::class);
+                $resolverWithSpreadMiddleware = $spreadMiddleware->wrap($resolverWithMiddleware);
+
+                return $resolverWithSpreadMiddleware($this->root, $this->args, $this->context, $this->resolveInfo);
             }
         );
 
@@ -257,17 +230,6 @@ class FieldFactory
             $type = $type->getWrappedType();
         }
 
-//        /** @var \Nuwave\Lighthouse\Schema\Extensions\ArgumentExtensions $extensions */
-//        $extensions = $input->config['lighthouse'] ?? null;
-
-//        if (
-//            $extensions->spread
-//            && $type instanceof InputObjectType
-//        ) {
-//            $this->pathsToSpread [] = $argumentPath;
-//        }
-
-        $directives = $this->directiveFactory->createAssociatedDirectivesOfType($astNode, ArgDirective::class);
         // Handle the argument itself. At this point, it can be wrapped
         // in a list or an input object
         $this->handleArgWithAssociatedDirectives($type, $astNode, $directives, $argumentPath);
@@ -394,11 +356,6 @@ class FieldFactory
     protected function setArgValue(array $argumentPath, $value): array
     {
         return Arr::set($this->args, implode('.', $argumentPath), $value);
-    }
-
-    protected function unsetArgValue(array $argumentPath): void
-    {
-        Arr::forget($this->args, implode('.', $argumentPath));
     }
 
     protected function argValue(array $argumentPath)